Enables -Wconversion for the simulator and fixes some warnings
Test: Run the sim, tests and load_chre.sh.

+++ platform/linux/system_time.cc
@@ -33,7 +33,11 @@ Nanoseconds SystemTime::getMonotonicTime() {
     return Nanoseconds(UINT64_MAX);
   }
 
-  return Seconds(timeNow.tv_sec) + Nanoseconds(timeNow.tv_nsec);
+  // The C++11 spec guarantees that tv_sec and tv_nsec only have values >= 0 and
+  // [0, 999999999]. It is safe to static cast these to their unsigned
+  // counterpart.
+  return Seconds(static_cast<uint64_t>(timeNow.tv_sec))
+      + Nanoseconds(static_cast<uint64_t>(timeNow.tv_nsec));
 }
 
 }  // namespace chre
